# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "江都 API",
    "version": "1.1.0",
    "description": "小程序+管理后台共用接口api"
  },
  "host": "localhost:4000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/api/tags/": {
      "get": {
        "tags": [
          "Tag"
        ],
        "summary": "获取Tag列表",
        "description": "",
        "parameters": [
          {
            "name": "TYPE",
            "in": "query",
            "description": "标签类别 PO/FI/TA...",
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ]
      },
      "put": {
        "tags": [
          "Tag"
        ],
        "summary": "修改Tag",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/tag_update"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tag"
        ],
        "summary": "新建Tag",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/tag_add"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tag"
        ],
        "summary": "删除Tag",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/tag_delete"
              }
            }
          }
        }
      }
    },
    "/api/talents/get": {
      "post": {
        "tags": [
          "Talent"
        ],
        "summary": "获取talent",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ]
      }
    },
    "/api/talents/": {
      "get": {
        "tags": [
          "Talent"
        ],
        "summary": "获取talent",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ]
      },
      "put": {
        "tags": [
          "Talent"
        ],
        "summary": "修改talent",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/talent_update"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Talent"
        ],
        "summary": "添加talent",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/user_register"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Talent"
        ],
        "summary": "删除talent",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/talent_delete"
              }
            }
          }
        }
      }
    },
    "/api/talents/tags": {
      "post": {
        "tags": [
          "Talent"
        ],
        "summary": "获取某个talent下全部tag",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/talents_tags"
              }
            }
          }
        }
      }
    },
    "/api/talents/allTags": {
      "post": {
        "tags": [
          "Talent"
        ],
        "summary": "应该是个废弃的接口了",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ]
      }
    },
    "/api/talents/addTags": {
      "post": {
        "tags": [
          "Talent"
        ],
        "summary": "给某个talent添加tag",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/talent_addTags"
              }
            }
          }
        }
      }
    },
    "/api/talents/deleteTags": {
      "post": {
        "tags": [
          "Talent"
        ],
        "summary": "给某个talent删除tag",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/talent_deleteTags"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "user_register": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "USER_NAME": {
              "type": "string",
              "example": "kevin"
            },
            "PASSWORD": {
              "type": "string",
              "example": "123"
            }
          }
        }
      }
    },
...
    "tag_add": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "TAG_ID": {
              "type": "string",
              "example": ""
            },
            "TAG_NAME": {
              "type": "string",
              "example": "dfd"
            },
            "TAG_VALUE": {
              "type": "string",
              "example": "fd"
            },
            "TYPE": {
              "type": "string",
              "example": "AS"
            },
            "DESCRIPTION": {
              "type": "string",
              "example": "fdfdfd"
            },
            "TAG_CATEGORY": {
              "type": "string",
              "example": "dfd"
            },
            "CREATED_AT": {
              "type": "string",
              "example": "2021-09-08T09:12:12.624Z"
            },
            "UPDATED_AT": {
              "type": "string",
              "example": "2021-09-08T09:12:12.624Z"
            },
            "USER_ID_USER_ID": {
              "type": "string",
              "example": "2200"
            }
          }
        },
        "token": {
          "type": "string",
          "example": "4836f6e1-2e8b-4b45-b903-9cf88d990901"
        }
      }
    },
    "tag_delete": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "text": {
              "type": "string",
              "example": "是否要删除dfd(ID: 8759ede4-5fa4-4d41-80f3-eaa2d0624cef)?"
            },
            "id": {
              "type": "string",
              "example": "8759ede4-5fa4-4d41-80f3-eaa2d0624cef"
            },
            "type": {
              "type": "string",
              "example": "tag"
            },
            "TAG_ID": {
              "type": "string",
              "example": "8759ede4-5fa4-4d41-80f3-eaa2d0624cef"
            }
          }
        }
      }
    },
    "tag_update": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "TAG_ID": {
              "type": "string",
              "example": "a694aa03-2a5e-4267-95cc-302921ddd21c"
            },
            "TAG_NAME": {
              "type": "string",
              "example": "efef"
            },
            "TAG_VALUE": {
              "type": "string",
              "example": "fefefddddd"
            },
            "TYPE": {
              "type": "string",
              "example": "FI"
            },
            "DESCRIPTION": {
              "type": "string",
              "example": "feefe"
            },
            "CREATED_AT": {
              "type": "string",
              "example": "2021-09-08"
            },
            "UPDATED_AT": {
              "type": "string",
              "example": "2021-09-08"
            },
            "USER_ID_USER_ID": {
              "type": "string",
              "example": "2200"
            },
            "USER_NAME": {
              "type": "string",
              "example": "null"
            },
            "TAG_CATEGORY": {
              "type": "string",
              "example": "ewfwff"
            }
          }
        }
      }
    },
    "talent_update": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "TALENT_ID": {
              "type": "string",
              "example": "13f9fcb8-0a36-11ec-85eb-90ccdf01187c"
            },
            "TALENT_NAME": {
              "type": "string",
              "example": "魏十五"
            },
            "AGE": {
              "type": "number",
              "example": 3
            },
            "SEX": {
              "type": "number",
              "example": 0
            },
            "POSITION": {
              "type": "string",
              "example": ""
            },
            "DEGREE": {
              "type": "number",
              "example": 1
            },
            "TYPE": {
              "type": "string",
              "example": ""
            },
            "UNIVERSITY": {
              "type": "string",
              "example": ""
            },
            "CAREER_YEAR": {
              "type": "number",
              "example": 2
            },
            "COMPANY_NAME": {
              "type": "string",
              "example": ""
            },
            "SALARY": {
              "type": "string",
              "example": "199k"
            },
            "AREA": {
              "type": "string",
              "example": ""
            },
            "RESUME_URL": {
              "type": "string",
              "example": ""
            },
            "CREATED_AT": {
              "type": "string",
              "example": "1970-01-01"
            },
            "UPDATED_AT": {
              "type": "string",
              "example": "1970-01-01"
            },
            "AVATAR": {
              "type": "string",
              "example": "https://s3.bmp.ovh/imgs/2021/09/31168b2e6cd39fb4.png"
            }
          }
        }
      }
    },
    "talents_tags": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "TALENT_ID_TALENT_ID": {
              "type": "string",
              "example": "020d921e-0a36-11ec-82f5-90ccdf01187c"
            }
          }
        }
      }
    },
    "talent_delete": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "text": {
              "type": "string",
              "example": "是否要删除undefined(ID: undefined)?"
            },
            "type": {
              "type": "string",
              "example": "policy"
            }
          }
        }
      }
    },
    "talent_addTags": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "TALENT_ID_TALENT_ID": {
              "type": "string",
              "example": "020d921e-0a36-11ec-82f5-90ccdf01187c"
            },
            "TAG_ID_TAG_ID": {
              "type": "string",
              "example": ""
            }
          }
        }
      }
    },
    "talent_deleteTags": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "TALENT_ID_TALENT_ID": {
              "type": "string",
              "example": "020d921e-0a36-11ec-82f5-90ccdf01187c"
            },
            "TAG_ID_TAG_ID": {
              "type": "string",
              "example": ""
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "JiangduJWT": {
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "JWT"
      }
    }
  }
}